# escrow-bounty/doc/swagger/escrow_bounty.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "application.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bounties": {
      "get": {
        "summary": "获取悬赏大厅列表",
        "operationId": "EscrowBountyService_ListBounties",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListBountiesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      },
      "post": {
        "summary": "雇主发布悬赏",
        "operationId": "EscrowBountyService_CreateBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateBountyRequest"
            }
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
...
    "/v1/bounties/{bountyId}": {
      "delete": {
        "summary": "雇主删除悬赏（仅 PENDING/FAILED 状态可删除）",
        "operationId": "EscrowBountyService_DeleteBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "bountyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
...
    "/v1/bounties/{id}": {
      "get": {
        "summary": "获取单个悬赏详情",
        "operationId": "EscrowBountyService_GetBounty",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetBountyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "EscrowBountyService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "pbBounty": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "employerUsername": {
          "type": "string"
        },
        "employerAccountId": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "rewardAmount": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "submissionText": {
          "type": "string"
        },
        "hunterUsername": {
          "type": "string",
          "title": "当前 ACCEPTED 的猎人用户名（无则空字符串）"
        },
        "employerAvatarUrl": {
          "type": "string"
        }
      }
    },
    "pbBountyApplication": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "bountyId": {
          "type": "string",
          "format": "int64"
        },
        "hunterUsername": {
          "type": "string"
        },
        "hunterAccountId": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "bounty": {
          "$ref": "#/definitions/pbBounty",
          "title": "关联悬赏信息（响应中附加）"
        },
        "hunterAvatarUrl": {
          "type": "string"
        }
      }
    },
...
    "pbCreateBountyRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "rewardAmount": {
          "type": "string",
          "format": "int64"
        },
        "employerAccountId": {
          "type": "string",
          "format": "int64"
        },
        "employerUsername": {
          "type": "string",
          "title": "可选；未提供则从 JWT token 中获取"
        },
        "deadlineTimestamp": {
          "type": "string",
          "format": "int64",
          "title": "截止时间（秒），可选"
        }
      },
      "title": "悬赏相关请求/响应"
    },
    "pbCreateBountyResponse": {
      "type": "object",
      "properties": {
        "bounty": {
          "$ref": "#/definitions/pbBounty"
        }
      }
    },
...
    "pbDeleteBountyResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "pbGetBountyResponse": {
      "type": "object",
      "properties": {
        "bounty": {
          "$ref": "#/definitions/pbBounty"
        },
        "applications": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbBountyApplication"
          }
        }
      }
    },
...
    "pbListBountiesResponse": {
      "type": "object",
      "properties": {
        "bounties": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbBounty"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}